Undo the integration with ca-certificates. Closes: #880419
The integration has had unintended side effects. All
the installed profiles will end up as trusted system
certificates regardless of the trust settings through
debconf. Since none of the IGTF certificates is meant
to be used for the typical tasks of verifying web sites
or e-mail users the integration is pointless.